requirements.txt
# Core dependencies
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain dependencies for agent functionality
langchain-core>=0.1.0
langchain-openai>=0.1.0
